Moved test to access suite
